Various tweaks
Spec & CHIP ingestion updates and local overrides.  There are no longer
validation errors on the final model except for a few enums & bitmaps defined
in CHIP that are not yet in the spec.

Changed attr/command/event IDs to hex to match spec.  Omit range bounds that
match the datatype bounds.