Move EndpointStructureLogger
Previously it was in util/ package.  This caused an import of just the util
package to import virtually the entire library.  This made codegen painful
because it uses some matter.js utility functions.  If the files to be
generated are not working then importing them would cause code gen to fail.

Now you can import from the util package and only the generic utilities will
load.