Matter 1.4
This includes application-level upgrades to Matter 1.4.  Upgrades the model, cluster definitions, endpoint types and behaviors.  Some protocol-level features are still TODO.

One notable change is that Basic Information "unique ID" is mandatory in 1.4.  We now generate one unconditionally unless provided as input.  This required tweaking current and deprecated codepaths and numerous tests but should be largely transparent to users.

Extends conformance and constraint parsers with new flags and syntax, plus a couple of bits that were used previously but we didn't support.  Some of these were complicated and consisted of ad-hoc syntax only used in one or two places...  But validation of the final model now sits at 100%.

Includes many upgrades to spec ingestion to handle special cases, mostly due to malformatting.  Replaces some hard-coded fixes with automatic detection/repair of errors that have become more prevalent as the spec grows.

Tweaks codegen for derived enums to generate union types that accept values from extended enums.  This appears to have been unnecessary because the specification was changed to include the redundant values, possibly because the equivalent fix was deemed infeasible in CHIP and/or alchemy.  However it does improve consistency so leaving in.

Includes a few minor CHIP test tweaks.  The new changes appear to have been transparent to CHIP but some test edits were no longer necessary and I fixed a few tests that had been broken previously.

Finally, includes a large chunk of code to handle ANSI text processing.  A bit of a rathole...  Came about because I wanted to make test error reporting more legible, which requires complex nested merging of styled text from multiple sources.  And I'd gotten tired of the mix of hand-coded styling and use of a 3rd-party library that has typing issues.  Over time we can migrate to this new code and drop the 3rd-party dependency, but will require some package reorganization because it currently lives in "tooling" which is not available everywhere.  So just approaching on an "as needed" basis for now.